Use socket2::Socket for UdpSocket
This includes a single breaking changes which involves multiple
UdpSocket methods: the receive methods now accept an uninitiliased
buffer ([MaybeUninit<u8>]), rather than an initialised one. This is the
case of the following methods:

* UdpSocket::recv_from
* UdpSocket::peek_from
* UdpSocket::recv
* UdpSocket::peek

This allows higher level libraries, such as Tokio or Heph, to use
uninitialised buffers if they want to. Note that the API guarantees not
to write any uninitialised data to the buffer, so it's not unsound to
pass `&[u8]` to the methods.

Finally this adds a debug assertions to UdpSocket::recv_from that checks
if the buffer passed isn't zero sized. Passing a zero sized buffer to
recv_from causes the OS to not initialise the source address (at least
on macOS), which in turn causes a misleading error to be returned saying
the address family is invalid. Before this change didn't check this and
likely returned an invalid IPv6 address.